requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
Django==5.0.3
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
